+++ src/Exception/NotFoundException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Obullo\Config\Exception;
-
-use RuntimeException;
-
-class NotFoundException extends RuntimeException
-{
-}